refactored home page to tsx and v3 API
