+++ src/components/CountryCard.tsx
@@ -2,23 +2,28 @@ import React from "react";
 import { Link } from "react-router-dom";
 import PropTypes from "prop-types";
 import "./countryCard.css";
+import { Country } from "../features/countries/countriesSlice";
 
-const CountryCard = ({ country }) => {
-  const { flags, name, population, region, capital, alpha3Code } = country;
+interface IProps {
+  country: Country;
+}
+
+const CountryCard = ({ country }: IProps) => {
+  const { flags, name, population, region, capital, cca3 } = country;
 
   const popFormat = new Intl.NumberFormat().format(population);
 
   return (
-    <Link to={`/country/${alpha3Code}`} className="country-card">
+    <Link to={`/country/${cca3}`} className="country-card">
       <img className="flag" src={flags.svg} alt={`flag of ${name}`} />
-      <h2 className="country-name">{name}</h2>
+      <h2 className="country-name">{name.common}</h2>
 
       <p className="country-basic-info">
         <b>Population:</b> {popFormat}
         <br />
         <b>Region:</b> {region}
         <br />
-        <b>Capital:</b> {capital}
+        <b>Capital:</b> {capital.map((capital) => capital)}
       </p>
     </Link>
   );
